Move ReAct prompt into agent runtime
